Make getTreeInfo an exported util function
The function will be useful to run tests in Webref
