+++ index.js
@@ -4,5 +4,6 @@ module.exports = {
   expandCrawlResult: require("./src/lib/util").expandCrawlResult,
   mergeCrawlResults: require("./src/lib/util").mergeCrawlResults,
   isLatestLevelThatPasses: require("./src/lib/util").isLatestLevelThatPasses,
+  getTreeInfo: require("./src/lib/util").getTreeInfo,
   postProcessor: require("./src/lib/post-processor")
 };
